--- Rules/IRules.cs
@@ -1,17 +1,0 @@
-using Table;
-using InfoGame;
-
-namespace Rules;
-
-#region StartGame
-
-public interface ICreateToken
-{
-    public int[,] Create(int n);
-}
-
-public interface IAssignTokenPlayer
-{
-}
-
-#endregion
