--- Rules/Rules.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <ItemGroup>
-    <ProjectReference Include="..\InfoGame\InfoGame.csproj" />
-    <ProjectReference Include="..\Table\Table.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Remove="GlobalUsings.cs" />
-  </ItemGroup>
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-</Project>
