+++ n-omino.sln
@@ -3,17 +3,19 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30114.105
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test", "Test\Test.csproj", "{049DB6C3-A79D-4549-A1C8-4B61BCC32B2A}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "n-ominoEngine", "n-ominoEngine", "{3C4AF272-12FA-4B58-BFCE-CCEEE12DB47A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Game", "Game\Game.csproj", "{212C33D6-D5DA-4409-BFD0-A4F28DEEEE0A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Game", "n-ominoEngine\Game\Game.csproj", "{C9DFC0AC-C84E-4D02-A01C-8D761A267B48}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Table", "Table\Table.csproj", "{D3FD35A2-301E-4D1B-9BCF-A2189ADBE65C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rules", "n-ominoEngine\Rules\Rules.csproj", "{009AF46F-E6E6-4C52-8DDA-4B5FECE6845F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Player", "Player\Player.csproj", "{007FF7B8-3890-4BB2-80D3-B15563FC34AA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InfoGame", "n-ominoEngine\InfoGame\InfoGame.csproj", "{D7AA48AF-E1F7-4C56-A940-6976A77C24C8}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rules", "Rules\Rules.csproj", "{31571934-56C2-46FA-8A32-3A706D14A7E6}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Table", "n-ominoEngine\Table\Table.csproj", "{D7C32684-CF44-42C6-A04D-A190144D5E41}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "InfoGame", "InfoGame\InfoGame.csproj", "{5B6FE2D1-C72B-4E3E-BA2B-9B61D7F9BF46}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Player", "n-ominoEngine\Player\Player.csproj", "{88FFF462-29C4-40A3-AACF-8856E63FD413}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "n-ominoServer", "n-ominoServer\n-ominoServer.csproj", "{EEF50E4B-BE1B-4E67-9983-C948E5ACD491}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -24,29 +26,36 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{049DB6C3-A79D-4549-A1C8-4B61BCC32B2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{049DB6C3-A79D-4549-A1C8-4B61BCC32B2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{049DB6C3-A79D-4549-A1C8-4B61BCC32B2A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{049DB6C3-A79D-4549-A1C8-4B61BCC32B2A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{212C33D6-D5DA-4409-BFD0-A4F28DEEEE0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{212C33D6-D5DA-4409-BFD0-A4F28DEEEE0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{212C33D6-D5DA-4409-BFD0-A4F28DEEEE0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{212C33D6-D5DA-4409-BFD0-A4F28DEEEE0A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D3FD35A2-301E-4D1B-9BCF-A2189ADBE65C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D3FD35A2-301E-4D1B-9BCF-A2189ADBE65C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D3FD35A2-301E-4D1B-9BCF-A2189ADBE65C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D3FD35A2-301E-4D1B-9BCF-A2189ADBE65C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{007FF7B8-3890-4BB2-80D3-B15563FC34AA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{007FF7B8-3890-4BB2-80D3-B15563FC34AA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{007FF7B8-3890-4BB2-80D3-B15563FC34AA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{007FF7B8-3890-4BB2-80D3-B15563FC34AA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{31571934-56C2-46FA-8A32-3A706D14A7E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{31571934-56C2-46FA-8A32-3A706D14A7E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{31571934-56C2-46FA-8A32-3A706D14A7E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{31571934-56C2-46FA-8A32-3A706D14A7E6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5B6FE2D1-C72B-4E3E-BA2B-9B61D7F9BF46}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5B6FE2D1-C72B-4E3E-BA2B-9B61D7F9BF46}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5B6FE2D1-C72B-4E3E-BA2B-9B61D7F9BF46}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5B6FE2D1-C72B-4E3E-BA2B-9B61D7F9BF46}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C9DFC0AC-C84E-4D02-A01C-8D761A267B48}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C9DFC0AC-C84E-4D02-A01C-8D761A267B48}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C9DFC0AC-C84E-4D02-A01C-8D761A267B48}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C9DFC0AC-C84E-4D02-A01C-8D761A267B48}.Release|Any CPU.Build.0 = Release|Any CPU
+		{009AF46F-E6E6-4C52-8DDA-4B5FECE6845F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{009AF46F-E6E6-4C52-8DDA-4B5FECE6845F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{009AF46F-E6E6-4C52-8DDA-4B5FECE6845F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{009AF46F-E6E6-4C52-8DDA-4B5FECE6845F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D7AA48AF-E1F7-4C56-A940-6976A77C24C8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D7AA48AF-E1F7-4C56-A940-6976A77C24C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D7AA48AF-E1F7-4C56-A940-6976A77C24C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D7AA48AF-E1F7-4C56-A940-6976A77C24C8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D7C32684-CF44-42C6-A04D-A190144D5E41}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D7C32684-CF44-42C6-A04D-A190144D5E41}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D7C32684-CF44-42C6-A04D-A190144D5E41}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D7C32684-CF44-42C6-A04D-A190144D5E41}.Release|Any CPU.Build.0 = Release|Any CPU
+		{88FFF462-29C4-40A3-AACF-8856E63FD413}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{88FFF462-29C4-40A3-AACF-8856E63FD413}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{88FFF462-29C4-40A3-AACF-8856E63FD413}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{88FFF462-29C4-40A3-AACF-8856E63FD413}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EEF50E4B-BE1B-4E67-9983-C948E5ACD491}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EEF50E4B-BE1B-4E67-9983-C948E5ACD491}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EEF50E4B-BE1B-4E67-9983-C948E5ACD491}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EEF50E4B-BE1B-4E67-9983-C948E5ACD491}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{C9DFC0AC-C84E-4D02-A01C-8D761A267B48} = {3C4AF272-12FA-4B58-BFCE-CCEEE12DB47A}
+		{009AF46F-E6E6-4C52-8DDA-4B5FECE6845F} = {3C4AF272-12FA-4B58-BFCE-CCEEE12DB47A}
+		{D7AA48AF-E1F7-4C56-A940-6976A77C24C8} = {3C4AF272-12FA-4B58-BFCE-CCEEE12DB47A}
+		{D7C32684-CF44-42C6-A04D-A190144D5E41} = {3C4AF272-12FA-4B58-BFCE-CCEEE12DB47A}
+		{88FFF462-29C4-40A3-AACF-8856E63FD413} = {3C4AF272-12FA-4B58-BFCE-CCEEE12DB47A}
 	EndGlobalSection
 EndGlobal
