+++ n-ominoEngine/Game/StartGame.cs
@@ -1,27 +1,27 @@
 using Table;
 
 namespace Game;
-/*
-public class StartGame<T>
+
+public class StartGame
 {
-    public List<Token<T>> Tokens { get; private set; }
+    public List<Token> Tokens { get; private set; }
     public StartGame(int players, int n, int k)
     {
         this.Tokens = CreateTokens(n, k);
     }
-    private List<Token<T>> CreateTokens(int n, int k)
+    private List<Token> CreateTokens(int n, int k)
     {
-        List<Token<T>> tokens = new List<Token<T>>();
+        List<Token> tokens = new List<Token>();
         Combination(0, 0, n, k, new int[n], tokens);
         return tokens;
     }
-    private void Combination(int ind, int start, int n, int k, int[] aux, List<Token<T>> tokens)
+    private void Combination(int ind, int start, int n, int k, int[] aux, List<Token> tokens)
     {
         if (ind == n)
         {
             int[] aux1 = new int[n];
             Array.Copy(aux, aux1, n);
-            tokens.Add(new Token<T>(aux1));
+            tokens.Add(new Token(aux1));
             return;
         }
         for (int i = start; i < k; i++)
@@ -30,4 +30,4 @@ public class StartGame<T>
             Combination(ind + 1, i, n, k, aux, tokens);
         }
     }
-}*/
+}
